+++ services/sso/src/controllers/dashboard/auth/login.ts
@@ -1,5 +1,5 @@
 import { Context, Middleware } from 'koa'
-import { client } from '../../../db/client'
+import db from '../../../db/knexClient'
 import { User } from '../../../schemas'
 import { UserLogin } from '../../../schemas/auth/loginSchema'
 import { UserRole, UserStatus } from '../../../schemas/users/userSchema'
@@ -12,14 +12,14 @@ import { appConfig } from '../../../config'
 export const dashboardLoginController: Middleware = async (ctx: Context) => {
   const { dni, password }: UserLogin = ctx.request.body
   const dniToUpperCase = dni.toUpperCase()
-  const userResult = await client.query(
-    'SELECT id, password, status, role FROM "user" WHERE dni = $1',
-    [dniToUpperCase]
-  )
-  const user = userResult.rows[0] as Pick<
-    User,
-    'id' | 'password' | 'status' | 'role'
-  >
+  const user:
+    | Pick<User, 'id' | 'password' | 'status' | 'role'>
+    | null
+    | undefined = await db<User>('user')
+    .select('id', 'password', 'status', 'role')
+    .where('dni', dniToUpperCase)
+    .first()
+
   if (!user) {
     throw new InvalidCredentials()
   }