+++ services/sso/src/controllers/dashboard/users/getMe.ts
@@ -1,19 +1,19 @@
 import { Context, Middleware } from 'koa'
 import { User } from '../../../schemas'
-import { client } from '../../../db/client'
 import { NotFoundError } from '../../../utils/errors'
+import { userManager } from '../../../db/managers/userManager'
 
 export const getMe: Middleware = async (ctx: Context) => {
   const { id } = ctx.state.user as Pick<User, 'id' | 'role'>
-  const queryResult = await client.query(
-    'SELECT dni, email, name, role FROM "user" WHERE id = $1',
-    [id]
-  )
 
-  const user = queryResult.rows[0]
+  const user = await userManager.findById(id, {
+    fields: ['dni', 'email', 'name', 'role'],
+  })
+
   if (!user) {
     throw new NotFoundError('User Not found')
   }
+
   ctx.status = 200
   ctx.body = user
 }