+++ services/sso/CHANGELOG.md
@@ -2,6 +2,48 @@ All notable changes to this project will be documented in this file.
 
 # Changelog
 
+### [1.38.0] - 2024-07-17
+
+### Changed
+
+- Migrated `POST /users/getME/` endpoint to Knex.
+
+## [1.37.0] - 2024-07-24
+
+### Changed
+
+- Updated `GET /itineraries` endpoint to Knex methods.
+
+## [1.36.0] - 2024-07-22
+
+### Changed
+
+- Updated `dashboard/user/getME` endpoint to Knex methods.
+
+### [1.35.0] - 2024-07-17
+
+### Changed
+
+- Migrated `POST /auth/register/` endpoint to Knex.
+
+## [1.34.3] - 2024-07-17
+
+### Added
+
+- Created `userManager` method in `userManager` for fetching user data by DNI.
+
+- Created `findByID` method in `userManager` for fetching user data by ID.
+
+### Updated
+
+- Migrated `POST /auth/login` endpoint to use the new methods and Knex
+
+## [1.34.2] - 2024-07-16
+
+### Changed
+
+- Migrate `POST /dashboard/users/` endpoint to knex and fix config database.
+
 ## [1.34.1] - 2024-06-19
 
 ### Added
@@ -12,7 +54,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
--  Implement knex initial config.
+- Implement knex initial config.
 
 ## [1.33.1] - 2024-06-19