+++ web/wiki/CHANGELOG.md
@@ -2,11 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.21.1] - 2024-07-24
+## [0.22.2] - 2024-07-25
 
 ### Fixed
 
-- Fix test of MyFavoritesList, MyResources, UserProfile, ResourceForm
+- Fixed useManageTopic hook and tests to pass tests related to `/api/v1/topics` endpoint
+- Changed openapiFetcher file to manage query string
+
+## [0.22.1] - 2024-07-25
+
+## Changed
+
+- Eliminated avatarId in tests that use useAuth mock
+
+## [0.22.0] - 2024-07-24
+
+### Deleted
+
+- Deleted AccountAdmin component and test
+- Deleted UsersManager component and test
 
 ## [0.21.0] - 2024-06-25